feat: controlar email a enviar en colas
